requirements.txt
librosa==0.9.2
numpy==1.26.3
ctranslate2
sentencepiece
faster-whisper==1.0.0
chardet==5.2.0
pytorch-lightning==2.2.0.post0
onnx==1.15.0
onnxruntime==1.17.0
onnxruntime-gpu==1.17.0
onnx2pytorch==0.4.1
soundfile==0.12.1
pydub==0.25.1
transformers==4.38.2
huggingface_hub==0.21.4
wordsegment==1.3.1
g2p-en==2.1.0
nltk==3.8.1
cn2an==0.5.22
pypinyin==0.51.0
jieba-fast==0.53
scipy==1.12.0
einops==0.7.0
tqdm==4.66.2
matplotlib==3.8.3
tensorboard==2.16.2
pytorch-lightning==2.2.1
PyYAML==6.0.1
gruut==2.3.4
typeguard==4.2.1
